feat: Implement OWM tie breaking
When two players have the same amount of wins, they are sorted by
their OWM scores. This is the total amount of wins their opponents
have. The idea is that the player who has a higher OWM score has
played against better opponents, and so they should win the tie.